requirements.txt
labelme==5.8.1
matplotlib==3.10.1
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.21.0
onnxsim==0.4.36
opencv_python==4.11.0.86
Pillow==11.2.1
scipy==1.15.2
thop==0.1.1.post2209072238
torch==2.6.0+cu126
torchsummary==1.5.1
tqdm==4.67.1
